fix redux state not register
